backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
replicate==0.28.0
openai==1.51.0
redis==5.1.1
celery==5.4.0
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2


